src/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0
yt-dlp>=2024.1.0
mutagen>=1.47.0
alembic>=1.13.0
python-dotenv>=1.0.0
python-multipart>=0.0.9

# Dev / testing
pytest>=8.0.0
httpx>=0.27.0
